Resume encoding without redoing all videos
If encoding was interrupted, current behavior is to restart encoding
all formats, including the ones that were already completed. This wastes
time.

This change tests for a given format is the target file already exists
with appropriate video duration, and if it is the case, we take it for
already completed.