Detect when an encoder process died
The current behavior when an encoder process crashes is to remain stuck
until someone change the encoding video status to queue in the database.

This changes introduce worker PID tracking. We add a worker_pid field
it the database, and we set it to current PID while running ffmpeg.

When checking status, we verify it the worker is still alive, and
we automatically set the video status to queue if it is not the case
anymore. That way, encoding can restart without manual database edition.

The queue can even be restarted if no user checks status by setting
up a crontab that checks status:
curl -so /dev/null http://localhost/avideo-encoder/view/run.php