Enable concurent encoding processing
Some systems have a lot of cores, it is a shame to leave them unused.
This changes let AVideo-Encoder perform concurent enconding.

The feature is controlled by global $concurent variable in the
configuration file. If absent, we fallback to 1, which means no
concurency. I experimented with $concurent set to 4 on a 4 core
systems with great benefit. Not all corner cases on status
update have been tested, though.