Send videos to streamer as soon as they are available
This change introduce a global option progressiveUpload so that
multi resolution runs send videos to the streamer as soon as they
have been individually encoded, instead of waiting for all of
them to be encoded.

I introduced an Upload class and a upload_queue database table
to keep track of what uploads have been done, so that the normal
process of sending all encoded videos can take over if a transfer
failed, while avoiding to send again the same file. This part may
be over-engineered, a global array may have been enough, thought it
would not have covered the situation where we resume encoding after
a crash.

Also remove a debug printf