Rollback unused change for progressive upload
When progressiveUpload is enabled, we send files with keepEncoding=1
POST paramter so that AVideo knows there are more videos to come,
and status should be k (active and encoding) instead of a (active).

I first planned to sendFile without a file at Encoder::delete() time
without keepEncoding=1 so that AVideo knows the video upload is over
and switch status to a. There was a bug (one missing parameter in
Encoder::sendFile(),  call) that caused it to be nilpotent, but the
behavior was still correct because Encoder:notifyVideoIsDone() also
causes to change the video status to a.

The logic to notify AVideo from Encoder::delete() by calling
Encoder:sendFile() without a file is therefore useless. This
patch removes it for the sake of code readability.