Add killWorkerOnDelete option to kill ffmpeg process on dequeue
We used to just let ffmpeg process run after a dequeue operation,
possibly wasting hours of CPU for big videos. This change introduces
a killWorkerOnDelete option so that the ffmpeg process is killed on
video dequeue.

In encoder_queue we have a new worker_ppid field which holds the
encodr apache process PID that was previosuly stored in worker_pid.
worker_pid now gets the ffmpeg PID.

The code to track the ffmpeg PID relies on the pcntl PHP extension.
Behavior hopefully fallback to what we have before this change is
the extension is absent.